requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary


# Test dependencies
pytest
pytest-cov
pytest-mock